mlops-stack/airflow/requirements.txt
joblib
mlflow==2.10.0
scikit-learn==1.0.2
numpy==1.24.4
xgboost
boto3
psycopg2-binary